fix: add inline edit wrapper component
