+++ packages/cloud-cognitive/src/components/InlineEdit/index.js
@@ -1,8 +1,1 @@
-/**
- * Copyright IBM Corp. 2022, 2022
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
 export { InlineEdit } from './InlineEdit';
