+++ packages/cloud-cognitive/src/components/InlineEditV1/_index.scss
@@ -5,4 +5,4 @@
 // LICENSE file in the root directory of this source tree.
 //
 
-@import './inline-edit';
+@import './inline-edit-v1';
