+++ packages/cloud-cognitive/src/components/InlineEditV1/_inline-edit-v1.scss
