+++ packages/cloud-cognitive/src/components/InlineEditV2/InlineEditV2.js
@@ -1,3 +1,10 @@
+/**
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 import React, { useState, useEffect, forwardRef, useRef } from 'react';
 import { Button } from 'carbon-components-react';
 import cx from 'classnames';
