+++ packages/cloud-cognitive/src/components/InlineEditV2/InlineEditV2.stories.js
@@ -1,9 +1,9 @@
-//
-// Copyright IBM Corp. 2020, 2021
-//
-// This source code is licensed under the Apache-2.0 license found in the
-// LICENSE file in the root directory of this source tree.
-//
+/**
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
 
 import React, { useState } from 'react';
 import {
@@ -11,6 +11,7 @@ import {
   prepareStory,
 } from '../../global/js/utils/story-helper';
 import { action } from '@storybook/addon-actions';
+import { InlineEdit } from '../InlineEdit/InlineEdit';
 import { InlineEditV2 } from '.';
 import mdx from './InlineEditV2.mdx';
 import styles from './_storybook-styles.scss';
@@ -41,6 +42,7 @@ const defaultProps = {
   readOnly: false,
   readOnlyLabel: 'This value is read only',
   saveLabel: 'Save',
+  v2: true,
   value: 'default',
 };
 
@@ -70,7 +72,7 @@ const Template = (args) => {
     onCancel,
   };
 
-  return <InlineEditV2 {...props} className="inline-edit-v2-example" />;
+  return <InlineEdit {...props} className="inline-edit-v2-example" />;
 };
 
 export const Default = prepareStory(Template, {
