+++ packages/cloud-cognitive/src/components/_index.scss
@@ -42,11 +42,12 @@
 @import './UserProfileImage/index';
 @import './EditSidePanel/index';
 @import './OptionsTile/index';
-@import './InlineEdit/index';
 @import './DataSpreadsheet/index';
 @import './Datagrid/index';
 @import './EditTearsheet/index';
 @import './EditTearsheetNarrow/index';
 @import './EditFullPage/index';
 @import './EditUpdateCards/index';
+
+@import './InlineEditV1/index';
 @import './InlineEditV2/index';
